feat: update agenda page
