+++ public/constant/agendas.ts
@@ -7,30 +7,52 @@ import stefan from "@/public/images/stefan.png";
 
 export type EventType = "hackathon" | "conference";
 
-export type AgendaType = {
-  time: string;
-  trackA?: {
-    topic: string;
-    duration: string;
-    speaker?: string;
-    title?: string;
-    src?: StaticImageData;
-  };
-  trackB?: {
-    topic: string;
-    duration: string;
-    speaker?: string;
-    title?: string;
-    src?: StaticImageData;
-  };
-};
-
 export const dates = {
   hackathon: [22, 23, 24],
   conference: [21, 22],
 };
 
-export const hackathonAgendas: { [key: number]: AgendaType[] } = {
+export type HackathonItemType = {
+  topic: string;
+  duration: string;
+  speaker?: string;
+  title?: string;
+  src?: StaticImageData;
+};
+
+export type HackathonAgenda = {
+  time: string;
+  trackA?: HackathonItemType;
+  trackB?: HackathonItemType;
+};
+
+export type ConferenceItemType = {
+  name: string;
+  company: string;
+  topic: string;
+  src?: StaticImageData | null;
+  duration: string;
+  panelOrganizations?: string[];
+  panelHost?: string;
+  panelists?: string[];
+};
+
+export type WorkshopItemType = {
+  holder: string;
+  title: string;
+  duration: string;
+  prerequisite?: string;
+  prerequisiteLink?: string;
+};
+
+export type ConferenceAgenda = {
+  time: string;
+  trackA?: ConferenceItemType;
+  trackB?: ConferenceItemType;
+  workshop?: WorkshopItemType;
+};
+
+export const hackathonAgendas: { [key: number]: HackathonAgenda[] } = {
   22: [
     {
       time: "6:30pm",
@@ -148,36 +170,7 @@ export const hackathonAgendas: { [key: number]: AgendaType[] } = {
 };
 
 export const conferenceAgendas: {
-  [key: number]: {
-    time: string;
-    trackA?: {
-      name: string;
-      company: string;
-      topic: string;
-      src?: StaticImageData | null;
-      duration: string;
-      panelOrganizations?: string[];
-      panelHost?: string;
-      panelists?: string[];
-    };
-    trackB?: {
-      name: string;
-      company: string;
-      topic: string;
-      src?: StaticImageData | null;
-      duration: string;
-      panelOrganizations?: string[];
-      panelHost?: string;
-      panelists?: string[];
-    };
-    workshop?: {
-      holder: string;
-      title: string;
-      duration: string;
-      prerequisite?: string;
-      prerequisiteLink?: string;
-    };
-  }[];
+  [key: number]: ConferenceAgenda[];
 } = {
   21: [
     {
